+++ .env
@@ -1,0 +1,6 @@
+NEXT_PUBLIC_API_KEY=AIzaSyAdjrZCa-2WG82dmHU1aII0g6cRdKYzoQg
+NEXT_PUBLIC_AUTH_DOMAIN=pet-venture-1777a.firebaseapp.com
+NEXT_PUBLIC_PROJECT_ID=pet-venture-1777a
+NEXT_PUBLIC_STORAGE_BUCKET=pet-venture-1777a.appspot.com
+NEXT_PUBLIC_MESSAGING_SENDER_ID=202804090837
+NEXT_PUBLIC_APP_ID=1:202804090837:web:69fcf8f98a1c2eefc20f5c
