--- src/app/.env
@@ -1,7 +1,0 @@
-API_KEY=AIzaSyAwcrHY5rIKNV57k9Bxj0pXKQRH1p7tUHs
-AUTH_DOMAIN=pet-venture.firebaseapp.com
-PROJECT_ID=pet-venture
-STORAGE_BUCKET=pet-venture.appspot.com
-MESSAGING_SENDER_ID=18245109239
-APP_ID=1:182451092395:web:9e8c2bdf9cf6d1fa31b0ee
-MEASUREMENT_ID=G-FVG72QFE24
